requirements.txt
python-dotenv==1.1.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
pandas==2.2.3
python-dotenv==1.1.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
pandas==2.2.3
requests==2.32.3
python-dotenv==1.1.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
pandas==2.2.3
requests==2.32.3
streamlit==1.44.1
plotly==6.0.1